# specs/akep/v0.1/openapi.yaml
openapi: 3.1.2
...
info:
  title: Agent Knowledge Exchange Protocol
  version: 0.1.0
  summary: Experimental HTTP binding for persistent, traceable Agent knowledge.
  description: |
    AKEP v0.1 Reader, Contributor, Evaluator, Curator, Publisher, and Federation operations.
    The capability document at `/.well-known/akep` supplies the actual base URL,
    supported profiles, authorization metadata, limits, and extensions.
  license:
    name: Project license pending
    identifier: LicenseRef-Pending
servers:
  - url: https://knowledge.example/akep/0.1
    description: Illustrative server; replace with discovery baseUrl
security:
  - oauth2: []
...
paths:
...
  /spaces/{spaceId}/attestations:
    post:
      tags: [Evaluator, Curator, Publisher]
      operationId: createAttestation
      summary: Persist an immutable review or validation statement
      description: |
        The issuer must equal the authenticated principal. Generic callers
        cannot mint benchmark-result statements; those are created only by a
        completed EvaluationRun. schema-validation and safety-scan are also
        reserved for checks actually executed by the node.
      security:
        - oauth2: [akep:evaluate]
        - oauth2: [akep:review]
        - oauth2: [akep:publish]
      parameters:
        - $ref: '#/components/parameters/AKEPVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/attestation.schema.json
      responses:
        '201':
          description: Immutable Attestation persisted idempotently
          headers:
            AKEP-Version:
              $ref: '#/components/headers/AKEPVersion'
            AKEP-Policy-Epoch:
              $ref: '#/components/headers/PolicyEpoch'
            Location:
              schema:
                type: string
                format: uri
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                const: 'private, no-store'
          content:
            application/json:
              schema:
                $ref: ./schemas/attestation.schema.json
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /spaces/{spaceId}/attestations/{attestationId}:
    get:
      tags: [Reader]
      operationId: fetchAttestation
      summary: Fetch an authorized immutable validation or review statement
      security:
        - oauth2: [akep:read]
      parameters:
        - $ref: '#/components/parameters/AKEPVersion'
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/Purpose'
        - name: attestationId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 2048
        - $ref: '#/components/parameters/Traceparent'
      responses:
        '200':
          description: Attestation statement; verify its referenced envelope when present
          headers:
            AKEP-Version:
              $ref: '#/components/headers/AKEPVersion'
            AKEP-Policy-Epoch:
              $ref: '#/components/headers/PolicyEpoch'
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                const: 'private, no-cache'
            Vary:
              schema:
                const: 'Authorization, AKEP-Purpose'
          content:
            application/json:
              schema:
                $ref: ./schemas/attestation.schema.json
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /evaluation-runs:
    post:
      tags: [Evaluator]
      operationId: createEvaluationRun
      summary: Record a completed evaluation and mint its benchmark Attestation
      description: |
        The service evaluates declared metrics against required and advisory
        thresholds, stores an immutable EvaluationRun, and atomically creates
        its benchmark-result Attestation. A failed required gate cannot satisfy
        publication quality policy.
      security:
        - oauth2: [akep:evaluate]
      parameters:
        - $ref: '#/components/parameters/AKEPVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/evaluation-run-request.schema.json
      responses:
        '201':
          description: Completed immutable EvaluationRun
          headers:
            AKEP-Version:
              $ref: '#/components/headers/AKEPVersion'
            AKEP-Policy-Epoch:
              $ref: '#/components/headers/PolicyEpoch'
            Location:
              schema:
                type: string
                format: uri
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                const: 'private, no-store'
          content:
            application/json:
              schema:
                $ref: ./schemas/evaluation-run.schema.json
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /evaluation-runs/{runId}:
    get:
      tags: [Reader]
      operationId: fetchEvaluationRun
      summary: Fetch an authorized immutable EvaluationRun
      security:
        - oauth2: [akep:read]
        - oauth2: [akep:review]
        - oauth2: [akep:publish]
      parameters:
        - $ref: '#/components/parameters/AKEPVersion'
        - $ref: '#/components/parameters/Purpose'
        - name: runId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 2048
        - $ref: '#/components/parameters/Traceparent'
      responses:
        '200':
          description: Immutable EvaluationRun and its quality-gate outcome
          headers:
            AKEP-Version:
              $ref: '#/components/headers/AKEPVersion'
            AKEP-Policy-Epoch:
              $ref: '#/components/headers/PolicyEpoch'
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                const: 'private, no-cache'
            Vary:
              schema:
                const: 'Authorization, AKEP-Purpose'
          content:
            application/json:
              schema:
                $ref: ./schemas/evaluation-run.schema.json
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  parameters:
    AKEPVersion:
      name: AKEP-Version
      in: header
      required: true
      schema:
        const: '0.1'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque key scoped to authenticated caller and operation
      schema:
        type: string
        minLength: 16
        maxLength: 255
    Traceparent:
      name: traceparent
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
    Purpose:
      name: AKEP-Purpose
      in: header
      required: true
      description: Requested use; it narrows but never grants token authority
      schema:
        type: string
        pattern: '^[a-z][a-z0-9._:-]{0,127}$'
...
    SpaceId:
      name: spaceId
      in: path
      required: true
      description: Percent-encoded canonical Space URI
      schema:
        type: string
        minLength: 1
        maxLength: 2048
...
  headers:
    AKEPVersion:
      description: Negotiated AKEP protocol version
      schema:
        const: '0.1'
    PolicyEpoch:
      description: Opaque epoch of the effective local policy snapshot
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
  responses:
...
    Problem:
      description: RFC 9457 problem details
      headers:
        AKEP-Version:
          $ref: '#/components/headers/AKEPVersion'
        Cache-Control:
          schema:
            const: 'private, no-store'
      content:
        application/problem+json:
          schema:
            $ref: ./schemas/problem.schema.json
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        Illustrative OAuth metadata. Clients must use the authorization server
        discovered from RFC 9728 protected resource metadata.
      flows:
        clientCredentials:
          tokenUrl: https://identity.example/oauth2/token
          scopes:
            akep:read: Read authorized immutable revisions
            akep:query: Query authorized knowledge
            akep:contribute: Submit candidate contributions
            akep:feedback: Submit usage evidence
            akep:review: Review and decide candidate contributions
            akep:publish: Publish or deprecate verified contributions
            akep:incident: Apply emergency revoke to a verified revoke contribution
            akep:erase: Apply privacy/legal erasure to a verified erase contribution
            akep:federate: Exchange signed federation changes